Clear the small warning classes and fix the defects they pointed at
The social editor accepted any position or level because its range tests
joined the bounds with &&, the high-level quest editor let negative class
numbers through, MOB_ADV_BLACKGUARD_MOUNT shared vnum 1234 with the basic
mount so the advanced mount (1236) was never summoned, and the track
command dereferenced a null victim when a charmed follower was told to
track an unknown name.

Unreachable branches are removed: a trap effect range check that could
never be true, the duergar weapon branch shadowed by WPT_DWARF's equal
value, an immortal usage text shadowed by the same level test, and a
duplicated staff level name. Identical branches are merged, lookups that
can fail get null guards, alloca and calloc(0) paths are gone, and
alignment-increasing casts go through void * or a copied header.

Printf, scanf, and strftime format attributes on 31 forwarding functions
let the compiler check their callers. That found 41 signedness mismatches,
cast to the printed type so NOWHERE still prints as -1 in the player file,
and 18 empty format strings.

-Wswitch-enum and -Wformat-nonliteral move to the analysis tier: the
switch sites are defaults over the event id enum, and the format sites
print from message tables by design.

+++ src/wilderness/spatial_core.c
@@ -614,7 +614,7 @@ spatial_direction_t spatial_calculate_direction(int observer_x, int observer_y, 
   else
     result = SPATIAL_DIR_UNKNOWN;
 
-  spatial_log("DEBUG: Final direction: %d", result);
+  spatial_log("DEBUG: Final direction: %d", (int)result);
   return result;
 }
